Updated lb code
